Fetch latest chagnes (#190)
* Add .DS_Store to .gitignore

This change ensures that macOS-specific .DS_Store files are ignored by Git. It helps prevent unnecessary or unintended tracking of these system files in the repository.

* Add copy task for Minecraft version 1.21.4 plugin

This change updates the Maven build process to include copying the plugin JAR file to the directory for Minecraft version 1.21.4. This ensures compatibility and proper deployment for the new version.